Move find_first/last tests to a separate file
